requirements.txt
Flask==3.0.0
numpy==1.26.0
tensorflow==2.18.0
Pillow==10.0.0
gunicorn==23.0.0
Werkzeug==3.0.1